Move: TicketCard 폴더 common 폴더로 이동(#4)
