+++ src/components/common/TicketCard/TicketCard.module.css
