+++ src/components/common/TicketCard/TicketCard.tsx
