+++ src/components/mainConcert/index.tsx
@@ -1,4 +1,3 @@
 import ConcertListSection from "./ConertListSection/ConcertListSection";
-import TicketCard from "./TicketCard/TicketCard";
 
-export { ConcertListSection, TicketCard };
+export { ConcertListSection };
